requirement.txt
requests==2.28.1
pandas==1.5.2
PyPDF2==2.12.1
pyTelegramBotAPI==4.8.0
reportlab==3.6.12
python-telegram-bot==13.15